refactor(recruit-board): RecruitBoard 객체 분리  - 봉사 정보를 VolunteerInfo 객체에서 관리
