+++ src/main/java/com/somemore/recruitboard/domain/RecruitBoard.java
@@ -7,14 +7,13 @@ import static lombok.AccessLevel.PROTECTED;
 
 import com.somemore.global.common.BaseEntity;
 import jakarta.persistence.Column;
+import jakarta.persistence.Embedded;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.Lob;
 import jakarta.persistence.Table;
-import java.time.Duration;
-import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.UUID;
 import lombok.Builder;
@@ -44,64 +43,30 @@ public class RecruitBoard extends BaseEntity {
     @Column(name = "content", nullable = false)
     private String content;
 
-    @Column(name = "region", nullable = false)
-    private String region;
-
-    @Column(name = "recruitment_count", nullable = false)
-    private Integer recruitmentCount;
-
-    @Column(name = "img_url", nullable = false)
-    private String imgUrl;
+    @Embedded
+    private VolunteerInfo volunteerInfo;
 
     @Enumerated(value = STRING)
     @Column(name = "recruit_status", nullable = false, length = 20)
     private RecruitStatus recruitStatus = RECRUITING;
 
-    @Column(name = "volunteer_start_date_time", nullable = false)
-    private LocalDateTime volunteerStartDateTime;
-
-    @Column(name = "volunteer_end_date_time", nullable = false)
-    private LocalDateTime volunteerEndDateTime;
-
-    @Enumerated(value = STRING)
-    @Column(name = "volunteer_type", nullable = false, length = 30)
-    private VolunteerType volunteerType;
-
-    @Column(name = "admitted", nullable = false)
-    private Boolean admitted;
+    @Column(name = "img_url", nullable = false)
+    private String imgUrl;
 
     @Builder
-    public RecruitBoard(UUID centerId, Long locationId, String title, String content, String region,
-        Integer recruitmentCount, String imgUrl, LocalDateTime volunteerStartDateTime,
-        LocalDateTime volunteerEndDateTime, VolunteerType volunteerType, Boolean admitted) {
-
-        validateVolunteerDateTime(volunteerStartDateTime, volunteerEndDateTime);
-
+    public RecruitBoard(UUID centerId, Long locationId, String title, String content,
+        VolunteerInfo volunteerInfo, String imgUrl) {
         this.centerId = centerId;
         this.locationId = locationId;
         this.title = title;
         this.content = content;
-        this.region = region;
-        this.recruitmentCount = recruitmentCount;
+        this.volunteerInfo = volunteerInfo;
         this.imgUrl = imgUrl;
-        this.volunteerStartDateTime = volunteerStartDateTime;
-        this.volunteerEndDateTime = volunteerEndDateTime;
-        this.volunteerType = volunteerType;
-        this.admitted = admitted;
     }
 
-    public LocalTime calculateVolunteerTime() {
-        Duration duration = Duration.between(volunteerStartDateTime, volunteerEndDateTime);
-
-        long hours = duration.toHours();
-        long minutes = duration.toMinutes() % 60;
-
-        return LocalTime.of((int) hours, (int) minutes);
+    public LocalTime getVolunteerHours() {
+        return volunteerInfo.calculateVolunteerTime();
     }
 
-    private void validateVolunteerDateTime(LocalDateTime startDateTime, LocalDateTime endDateTime) {
-        if (endDateTime.isEqual(startDateTime) || endDateTime.isBefore(startDateTime)) {
-            throw new IllegalArgumentException("종료 시간은 시작 시간보다 이후여야 합니다.");
-        }
-    }
+
 }
